made new component for profile button
